+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - `cli/list`: Show plugin braces only there's plugin output
 - `lib/tman`: Change return value of function tman_deinit()
 - `lib/unit`: Change return value of function unit_delpgn()
+- `lib/tree`: Change return value of function tree_free()
 
 ### Fixed